+++ packages/admin-e2e-tests/CHANGELOG.md
@@ -1,6 +1,6 @@
 # Unreleased
 
--   Add test cases for home screent tasklist and activity panels. #7509
+-   Add test cases for the home screen tasklist and activity panels. #7509
 
 # 0.1.0